fixup! Let a global implement another global's callback
