+++ internal/compiler/passes.rs
@@ -73,45 +73,6 @@ pub fn ignore_debug_hooks(expr: &Expression) -> &Expression {
     }
 }
 
-/// Return the alias node of a `callback foo <=> ...;` declaration, if `prop` is one.
-///
-/// This lives on the callback declaration itself, which is where the alias of a
-/// global callback that also has a handler ends up (the handler takes the binding
-/// expression slot).
-pub fn callback_alias_declaration_node(
-    elem: &crate::object_tree::Element,
-    prop: &str,
-) -> Option<crate::parser::syntax_nodes::TwoWayBinding> {
-    elem.property_declarations
-        .get(prop)
-        .and_then(|d| d.node.clone())
-        .and_then(crate::parser::syntax_nodes::CallbackDeclaration::new)
-        .and_then(|cb| cb.TwoWayBinding())
-}
-
-/// Return the two-way-binding syntax node of a `<=>` alias for the given property, if any.
-///
-/// Usually the alias is the binding's own (uncompiled) expression. But a global
-/// callback may both alias another global's callback (`callback foo <=> Other.foo;`)
-/// and provide a handler (`foo => { ... }`): the handler then occupies the binding
-/// expression slot, so the alias node lives on the callback declaration instead.
-pub fn two_way_binding_node(
-    elem: &crate::object_tree::Element,
-    prop: &str,
-) -> Option<crate::parser::syntax_nodes::TwoWayBinding> {
-    use crate::parser::syntax_nodes;
-    if let Some(binding) = elem.bindings.get(prop) {
-        if let Ok(b) = binding.try_borrow() {
-            if let Expression::Uncompiled(node) = ignore_debug_hooks(&b.expression) {
-                if let Some(twb) = syntax_nodes::TwoWayBinding::new(node.clone()) {
-                    return Some(twb);
-                }
-            }
-        }
-    }
-    callback_alias_declaration_node(elem, prop)
-}
-
 pub async fn run_passes(
     doc: &mut crate::object_tree::Document,
     type_loader: &mut crate::typeloader::TypeLoader,
